# Optimized SHA1 and SHA256