First, some background. Both ARM and Intel are releasing documentation that they plan to include instruction set extensions for SHA1 and SHA256 sometime in the near future. Intel has [some documentation](https://software.intel.com/en-us/articles/intel-sha-extensions) about it, and I wrote [a blog entry](http://straymindcough.blogspot.com/2015/01/hardware-accelerated-sha-1.html) about it, too.

Second, LLVM won't lower these intrinsics yet, so I started writing some functions to emulate the unusable intrinsics, and implemented SHA1 in terms of those emulated intrinsic functions, and it turned out to be quite fast! So I tried the same thing with SHA256 (which will have HW instructions at some point), and SHA512 (which I couldn't find any instruction designs for, so I made up my own).

I believe that one reason these implementations are faster is possibly because it uses packed addition and packed xor instead of scalar arithmetic. I haven't tested it on any architectures other than x86-64, so it may be that it ends up being slower on other architectures. I kept the current code in a function with the same name with a "`_safe`" appended to it. So if it ends up being slower on ARM, for example, we, or users of the library, can switch the implementations, if needed.