@DaGenix 2015-02-14
I did a little light benchmarking on my phone - a Samsung Galaxy Nexus - so, an older phone. My first finding - it looks pretty unlikely that anyone is currently using rust-crypto on ARM - it didn't compile. I've fixed that. Next, the time library on crates.io didn't compile either - I work around that and I'm going to open an issue for that. After working through that, it looks like the benchmarks are fairly unstable - I couldn't get the benchmark program to run more than 3-4 tests before hanging and I have no idea why. Anyway, stitching together results from multiple runs, I got the following, very rough, benchmarks:

Current master:

```
test sha1::bench::sha1_10                          ... bench:       482 ns/iter (+/- 46) = 20 MB/s
test sha1::bench::sha1_1k                          ... bench:     36430 ns/iter (+/- 5150) = 28 MB/s
test sha1::bench::sha1_64k                         ... bench:   2339935 ns/iter (+/- 78658) = 27 MB/s
test sha2::bench::sha256_10                        ... bench:       667 ns/iter (+/- 12) = 14 MB/s
test sha2::bench::sha256_1k                        ... bench:     55313 ns/iter (+/- 972) = 18 MB/s
test sha2::bench::sha256_64k                       ... bench:   3594970 ns/iter (+/- 46081) = 18 MB/s
test sha2::bench::sha512_10                        ... bench:      1084 ns/iter (+/- 24) = 9 MB/s
test sha2::bench::sha512_1k                        ... bench:     97656 ns/iter (+/- 591) = 10 MB/s
test sha2::bench::sha512_64k                       ... bench:   6304931 ns/iter (+/- 55541) = 10 MB/s
```

With your changes:

```
test sha1::bench::sha1_10                          ... bench:       373 ns/iter (+/- 2) = 26 MB/s
test sha1::bench::sha1_1k                          ... bench:     24735 ns/iter (+/- 24) = 41 MB/s
test sha1::bench::sha1_64k                         ... bench:   1629638 ns/iter (+/- 6103) = 40 MB/s
test sha2::bench::sha256_10                        ... bench:       536 ns/iter (+/- 24) = 18 MB/s
test sha2::bench::sha256_1k                        ... bench:     41580 ns/iter (+/- 763) = 24 MB/s
test sha2::bench::sha256_64k                       ... bench:   2661132 ns/iter (+/- 155945) = 24 MB/s
test sha2::bench::sha512_10                        ... bench:      1084 ns/iter (+/- 12) = 9 MB/s
test sha2::bench::sha512_1k                        ... bench:     95748 ns/iter (+/- 763) = 10 MB/s
test sha2::bench::sha512_64k                       ... bench:   6115722 ns/iter (+/- 37841) = 10 MB/s
```

So, it looks like sha1 is faster on my phone and sha2 is unchanged.

I haven't had time to really dig into your changes. However, an initial question - are the *_safe variants ever called? It doesn't look to me like they are.