@andydude 2015-02-15
No, the _safe variants are just the previous code with some minor changes to support switching between the two. I didn't have enough information to know which would be faster, but it looks like the previous code would be better on ARM.