@DaGenix 2015-02-17
Sorry for not getting back to you earlier - I was feeling a bit under the weather this weekend. I keep running into problems benchmarking on my devices. I do wonder if there is an issue in the benchmarking code under Android. Anyway, I tried again on another device and here is what I see on ARM:

Current code:
```
test sha1::bench::sha1_1k                          ... bench:     14877 ns/iter (+/- 191) = 68 MB/s
test sha2::bench::sha256_1k                        ... bench:     23078 ns/iter (+/- 191) = 44 MB/s
test sha2::bench::sha512_1k                        ... bench:     35095 ns/iter (+/- 763) = 29 MB/s
```

With your code, but without NEON enabled (I think):
```
test sha1::bench::sha1_1k                          ... bench:     11825 ns/iter (+/- 382) = 86 MB/s
test sha2::bench::sha256_1k                        ... bench:     19645 ns/iter (+/- 243) = 52 MB/s
test sha2::bench::sha512_1k                        ... bench:     93078 ns/iter (+/- 2365) = 11 MB/s
```

With your code with NEON enabled (I think):
```
test sha1::bench::sha1_1k                          ... bench:     13542 ns/iter (+/- 5197) = 75 MB/s
test sha2::bench::sha256_1k                        ... bench:     18463 ns/iter (+/- 1434) = 55 MB/s
test sha2::bench::sha512_1k                        ... bench:     69427 ns/iter (+/- 1526) = 14 MB/s
```

So, as best as I can tell, it looks like this code performs better - except for SHA512. I don't know why that is. Anyway, I'm fairly certain that no one is actually using rust-crypto on ARM. So, even if there are performance regression on ARM, I don't believe that should hold up merging these changes. I'd prefer not to have unused code sitting around - could you remove the old, unused SHA1 / SHA2 code and the comment saying that the appropriate versions are picked at runtime (unless you'd prefer to use that code in some way).

I'll leave a few more comments inline with the patches.

Thanks again for the PR!