@DaGenix 2015-02-17
Also, what would you think of giving the functions somewhat more descriptive names. Right now, they are all named after the intrinsics that they implement - which does make sense - but its a bit tricky to read if you aren't familiar with them (as I'm not). If you think changing the names would make the code harder to read, please let me know - just something I wanted to discuss.