@andydude 2015-02-18
Ah, descriptive names. I actually prefer the ARM names to the x86-64 names, because if I were to name them Rust-style, then they would probably be named:

<table>
    <tr>
        <th>ARM</th>
        <th>x86</th>
        <th>Rust?</th>
    </tr>
    <tr>
        <td>sha1su0</td>
        <td>sha1msg1</td>
        <td rowspan="2">sha1_schedule_x4 (4 rounds of message schedule update)</td>
    </tr>
    <tr>
        <td>sha1su1</td>
        <td>sha1msg2</td>
    </tr>
    <tr>
        <td>sha1c</td>
        <td>sha1rnds4 0</td>
        <td rowspan="5">sha1_digest_round_x4 (4 rounds of message block digest)</td>
    </tr>
    <tr>
        <td>sha1p</td>
        <td>sha1rnds4 1</td>
    </tr>
    <tr>
        <td>sha1m</td>
        <td>sha1rnds4 2</td>
    </tr>
    <tr>
        <td>sha1p</td>
        <td>sha1rnds4 3</td>
    </tr>
    <tr>
        <td>sha1h</td>
        <td>sha1nexte</td>
    </tr>
</table>

The function signatures would probably be the same as the macros used in the implementation.
And if you noticed, the SHA-256 and SHA-512 implementations also use similar macros, and so
I would probably use similar names for those:

* `sha256_schedule_x4`
* `sha256_digest_round_x4`
* `sha512_schedule_x4`
* `sha512_digest_round_x4`