@DaGenix 2015-02-18
On my computer, I'm seeing SHA1 go from 250MB/s to almost 400MB/s which is very cool. SHA2 also has nice speed boosts.

Thanks!